requirements.txt
streamlit
plotly
pyyaml
sentence-transformers